+++ api/src/controllers/referent.js
@@ -6,6 +6,7 @@ const jwt = require("jsonwebtoken");
 const mime = require("mime-types");
 const FileType = require("file-type");
 const Joi = require("joi");
+const NodeClam = require("clamscan");
 
 const ReferentModel = require("../models/referent");
 const YoungModel = require("../models/young");
@@ -31,7 +32,6 @@ const {
   updatePlacesSessionPhase1,
   updatePlacesBus,
   signinLimiter,
-  //  assignNextYoungFromWaitingList,
   ERRORS,
   isYoung,
   inSevenDays,
@@ -311,10 +311,18 @@ router.put("/young/:id", passport.authenticate("referent", { session: false, fai
       if (young.statusPhase3 !== "VALIDATED") newYoung.statusPhase3 = "WITHDRAWN";
     }
 
-    // if withdrawn from phase1 -> run the script that find a replacement for this young
-    if (newYoung.statusPhase1 === "WITHDRAWN" && ["AFFECTED", "WAITING_ACCEPTATION"].includes(young.statusPhase1) && young.cohesionCenterId) {
-      // disable the 08 jun 21
-      // await assignNextYoungFromWaitingList(young);
+    if (newYoung?.department && young?.department && newYoung?.department !== young?.department) {
+      const referents = await ReferentModel.find({ department: newYoung.department, role: ROLES.REFERENT_DEPARTMENT });
+      for (let referent of referents) {
+        await sendTemplate(SENDINBLUE_TEMPLATES.young.DEPARTMENT_CHANGE, {
+          emailTo: [{ name: `${referent.firstName} ${referent.lastName}`, email: referent.email }],
+          params: {
+            youngFirstName: young.firstName,
+            youngLastName: young.lastName,
+            cta: `${config.ADMIN_URL}/volontaire/${young._id}`,
+          },
+        });
+      }
     }
 
     if (newYoung.cohesionStayPresence === "true" && young.cohesionStayPresence !== "true") {
@@ -344,20 +352,6 @@ router.put("/young/:id", passport.authenticate("referent", { session: false, fai
       if (qpv === true) newYoung.qpv = "true";
       else if (qpv === false) newYoung.qpv = "false";
       else newYoung.qpv = "";
-    }
-
-    if (newYoung.department !== young.department) {
-      const referents = await ReferentModel.find({ department: newYoung.department, role: ROLES.REFERENT_DEPARTMENT });
-      for (let referent of referents) {
-        await sendTemplate(SENDINBLUE_TEMPLATES.young.DEPARTMENT_CHANGE, {
-          emailTo: [{ name: `${referent.firstName} ${referent.lastName}`, email: referent.email }],
-          params: {
-            youngFirstName: newYoung.firstName,
-            youngLastName: newYoung.lastName,
-            cta: `${config.ADMIN_URL}/volontaire/${newYoung._id}`,
-          },
-        });
-      }
     }
 
     // Check quartier prioritaires.
@@ -598,11 +592,13 @@ router.post("/file/:key", passport.authenticate("referent", { session: false, fa
           Joi.object({
             name: Joi.string().required(),
             data: Joi.binary().required(),
+            tempFilePath: Joi.string().allow("").optional(),
           }).unknown(),
           Joi.array().items(
             Joi.object({
               name: Joi.string().required(),
               data: Joi.binary().required(),
+              tempFilePath: Joi.string().allow("").optional(),
             }).unknown(),
           ),
         ),
@@ -618,7 +614,17 @@ router.post("/file/:key", passport.authenticate("referent", { session: false, fa
       if (Array.isArray(currentFile)) {
         currentFile = currentFile[currentFile.length - 1];
       }
-      const { name, data } = currentFile;
+      const { name, data, tempFilePath } = currentFile;
+
+      if (config.ENVIRONMENT === "staging") {
+        const clamscan = await new NodeClam().init({
+          removeInfected: true,
+        });
+        const { isInfected } = await clamscan.isInfected(tempFilePath);
+        if (isInfected) {
+          return res.status(400).send({ ok: false, code: ERRORS.INVALID_PARAMS, error: "File is infected" });
+        }
+      }
 
       const encryptedBuffer = encrypt(data);
       const resultingFile = { mimetype: "image/png", encoding: "7bit", data: encryptedBuffer };