+++ api/src/index.js
@@ -48,7 +48,12 @@ app.use(bodyParser.urlencoded({ extended: true }));
 
 require("./crons");
 app.use(cookieParser());
-app.use(fileUpload({ limits: { fileSize: 10 * 1024 * 1024 } })); // 10 Mo
+if (ENVIRONMENT !== "staging") {
+  app.use(fileUpload({ limits: { fileSize: 10 * 1024 * 1024 } })); // 10 Mo
+} else {
+  app.use(fileUpload({ limits: { fileSize: 10 * 1024 * 1024 }, useTempFiles: true, tempFileDir: "/tmp/" }));
+}
+
 app.use(express.static(__dirname + "/../public"));
 
 app.use(passport.initialize());
@@ -78,6 +83,12 @@ app.use("/support-center/user", require("./controllers/supportUser"));
 app.use("/support-center/ticket", require("./controllers/supportTicket"));
 app.use("/support-center/knowledge-base", require("./controllers/knowledgeBase"));
 app.use("/signin", require("./controllers/signin"));
+
+//services
+if (ENVIRONMENT !== "production") {
+  app.use("/jeveuxaider", require("./services/jeveuxaider"));
+}
+
 app.use(handleError);
 
 app.get("/", async (req, res) => {