+++ api/src/utils/index.js
@@ -3,7 +3,6 @@ const https = require("https");
 const http = require("http");
 const passwordValidator = require("password-validator");
 const YoungModel = require("../models/young");
-const CohesionCenterModel = require("../models/cohesionCenter");
 const MeetingPointModel = require("../models/meetingPoint");
 const ApplicationModel = require("../models/application");
 const ReferentModel = require("../models/referent");
@@ -12,8 +11,7 @@ const { sendEmail, sendTemplate } = require("../sendinblue");
 const path = require("path");
 const fs = require("fs");
 const rateLimit = require("express-rate-limit");
-const sendinblue = require("../sendinblue");
-const { ADMIN_URL, APP_URL } = require("../config");
+const { APP_URL } = require("../config");
 const {
   CELLAR_ENDPOINT,
   CELLAR_KEYID,
@@ -25,8 +23,8 @@ const {
   API_ASSOCIATION_CELLAR_KEYID,
   API_ASSOCIATION_CELLAR_KEYSECRET,
 } = require("../config");
-const { ROLES } = require("snu-lib/roles");
 const { YOUNG_STATUS_PHASE2, SENDINBLUE_TEMPLATES, YOUNG_STATUS } = require("snu-lib/constants");
+const { getQPV, getDensity } = require("../geo");
 
 // Set the number of requests allowed to 15 in a 1 hour window
 const signinLimiter = rateLimit({
@@ -135,7 +133,7 @@ function getSignedUrlForApiAssociation(path) {
 }
 
 function fileExist(url) {
-  return new Promise((resolve, reject) => {
+  return new Promise((resolve) => {
     getReq(url, (resp) => {
       if (resp.statusCode === 200) return resolve(true);
       return resolve(false);
@@ -281,32 +279,6 @@ const updatePlacesBus = async (bus) => {
   return bus;
 };
 
-const sendAutoAffectationMail = async (nextYoung, center) => {
-  // Send mail.
-  const cc = [];
-  if (nextYoung.parent1Email) cc.push({ email: nextYoung.parent1Email });
-  if (nextYoung.parent2Email) cc.push({ email: nextYoung.parent2Email });
-  await sendEmail(
-    {
-      name: `${nextYoung.firstName} ${nextYoung.lastName}`,
-      email: nextYoung.email,
-    },
-    "Une place dans le séjour de cohésion SNU 2021 s’est libérée !",
-    fs
-      .readFileSync(path.resolve(__dirname, "./templates/autoAffectation.html"))
-      .toString()
-      .replace(/{{firstName}}/, nextYoung.firstName)
-      .replace(/{{lastName}}/, nextYoung.lastName)
-      .replace(/{{centerName}}/, center.name)
-      .replace(/{{centerAddress}}/, center.address + " " + center.zip + " " + center.city)
-      .replace(/{{centerDepartement}}/, center.department)
-      .replace(/{{ctaAccept}}/, "https://moncompte.snu.gouv.fr/auth/login?redirect=phase1")
-      .replace(/{{ctaDocuments}}/, "https://moncompte.snu.gouv.fr/auth/login?redirect=phase1")
-      .replace(/{{ctaWithdraw}}/, "https://moncompte.snu.gouv.fr/auth/login?redirect=phase1"),
-    { cc },
-  );
-};
-
 const sendAutoCancelMeetingPoint = async (young) => {
   const cc = [];
   if (young.parent1Email) cc.push({ email: young.parent1Email });
@@ -327,61 +299,6 @@ const sendAutoCancelMeetingPoint = async (young) => {
   );
 };
 
-const sendAutoAffectationNotFoundMails = async (to, young, center) => {
-  // Send mail.
-  await sendEmail(
-    {
-      name: `${to.firstName} ${to.lastName}`,
-      email: to.email,
-    },
-    "Une place s'est libérée dans l'un de vos centres de séjour SNU",
-    fs
-      .readFileSync(path.resolve(__dirname, "./templates/autoAffectationNotFound.html"))
-      .toString()
-      .replace(/{{firstName}}/, to.firstName)
-      .replace(/{{lastName}}/, to.lastName)
-      .replace(/{{youngFirstName}}/, young.firstName)
-      .replace(/{{youngLastName}}/, young.lastName)
-      .replace(/{{centerName}}/, center.name)
-      .replace(/{{cta}}/, `${ADMIN_URL}/auth?redirect=centre/${center._id}/affectation`),
-  );
-};
-
-const assignNextYoungFromWaitingList = async (young) => {
-  const nextYoung = await getYoungFromWaitingList(young);
-  if (!nextYoung) {
-    //notify referents & admin
-    console.log(`no replacement found for young ${young._id} in center ${young.cohesionCenterId}`);
-
-    const center = await CohesionCenterModel.findById(young.cohesionCenterId);
-    if (!center) return null;
-    let to = await ReferentModel.find({ role: ROLES.ADMIN, email: { $in: ["youssef.tahiri@education.gouv.fr", "nicolas.roy@recherche.gouv.fr"] } });
-    to = to.concat(await ReferentModel.find({ role: ROLES.REFERENT_REGION, region: center.region }));
-    for (let i = 0; i < to.length; i++) {
-      await sendAutoAffectationNotFoundMails(to[i], young, center);
-    }
-  } else {
-    // Notify young & modify statusPhase1
-    console.log("replacement found", nextYoung._id);
-
-    // Activate waiting accepation and 48h cron
-    nextYoung.set({ status: "VALIDATED", statusPhase1: "WAITING_ACCEPTATION", autoAffectationPhase1ExpiresAt: Date.now() + 60 * 1000 * 60 * 48 });
-    await nextYoung.save();
-    await sendinblue.sync(nextYoung, "young");
-
-    const center = await CohesionCenterModel.findById(nextYoung.cohesionCenterId);
-    await sendAutoAffectationMail(nextYoung, center);
-
-    //remove the young from the waiting list
-    if (center?.waitingList?.indexOf(nextYoung._id) !== -1) {
-      console.log(`remove young ${nextYoung._id} from waiting_list of ${nextYoung.cohesionCenterId}`);
-      const i = center.waitingList.indexOf(nextYoung._id);
-      center.waitingList.splice(i, 1);
-      await center.save();
-    }
-  }
-};
-
 // pourrait être utile un jour
 
 // const assignYoungToWaitingList = async (young, newCohort) => {
@@ -399,25 +316,6 @@ const assignNextYoungFromWaitingList = async (young) => {
 //   }
 
 // }
-
-const getYoungFromWaitingList = async (young) => {
-  try {
-    if (!young || !young.cohesionCenterId) return null;
-    const center = await CohesionCenterModel.findById(young.cohesionCenterId);
-    if (!center) return null;
-    let res = null;
-    for (let i = 0; i < center.waitingList?.length; i++) {
-      const tempYoung = await YoungModel.findById(center.waitingList[i]);
-      if (tempYoung.statusPhase1 === "WAITING_LIST" && tempYoung.department === young.department && tempYoung.gender === young.gender) {
-        res = tempYoung;
-        break;
-      }
-    }
-    return res;
-  } catch (e) {
-    console.log(e);
-  }
-};
 
 async function updateYoungPhase2Hours(young) {
   const applications = await ApplicationModel.find({
@@ -563,6 +461,38 @@ const isObjectKeysIsEqual = (object, newObject, keys) => {
   return true;
 };
 
+async function inscriptionCheck(value, young, req) {
+  // Check quartier prioritaires.
+  if (value.zip && value.city && value.address) {
+    const qpv = await getQPV(value.zip, value.city, value.address);
+    if (qpv === true) young.set({ qpv: "true" });
+    else if (qpv === false) young.set({ qpv: "false" });
+    else young.set({ qpv: "" });
+    await young.save({ fromUser: req.user });
+  }
+
+  // Check quartier prioritaires.
+  if (value.cityCode) {
+    const populationDensity = await getDensity(value.cityCode);
+    young.set({ populationDensity });
+    await young.save({ fromUser: req.user });
+  }
+
+  // if withdrawn, cascade withdrawn on every status
+  if (young.status === "WITHDRAWN" && (young.statusPhase1 !== "WITHDRAWN" || young.statusPhase2 !== "WITHDRAWN" || young.statusPhase3 !== "WITHDRAWN")) {
+    if (young.statusPhase1 !== "DONE") young.set({ statusPhase1: "WITHDRAWN" });
+    if (young.statusPhase2 !== "VALIDATED") young.set({ statusPhase2: "WITHDRAWN" });
+    if (young.statusPhase3 !== "VALIDATED") young.set({ statusPhase3: "WITHDRAWN" });
+    await young.save({ fromUser: req.user });
+  }
+
+  // if they had a cohesion center, we check if we need to update the places taken / left
+  if (young.sessionPhase1Id) {
+    const sessionPhase1 = await SessionPhase1.findById(young.sessionPhase1Id);
+    if (sessionPhase1) await updatePlacesSessionPhase1(sessionPhase1);
+  }
+}
+
 const ERRORS = {
   SERVER_ERROR: "SERVER_ERROR",
   NOT_FOUND: "NOT_FOUND",
@@ -581,6 +511,7 @@ const ERRORS = {
   INVALID_BODY: "INVALID_BODY",
   INVALID_PARAMS: "INVALID_PARAMS",
   EMAIL_OR_PASSWORD_INVALID: "EMAIL_OR_PASSWORD_INVALID",
+  EMAIL_OR_TOKEN_INVALID: "EMAIL_OR_TOKEN_INVALID",
   PASSWORD_INVALID: "PASSWORD_INVALID",
   EMAIL_INVALID: "EMAIL_INVALID",
   EMAIL_AND_PASSWORD_REQUIRED: "EMAIL_AND_PASSWORD_REQUIRED",
@@ -589,6 +520,33 @@ const ERRORS = {
   USER_NOT_EXISTS: "USER_NOT_EXISTS",
   NEW_PASSWORD_IDENTICAL_PASSWORD: "NEW_PASSWORD_IDENTICAL_PASSWORD",
   INVALID_IP: "INVALID_IP",
+};
+
+const YOUNG_SITUATIONS = {
+  GENERAL_SCHOOL: "GENERAL_SCHOOL",
+  PROFESSIONAL_SCHOOL: "PROFESSIONAL_SCHOOL",
+  AGRICULTURAL_SCHOOL: "AGRICULTURAL_SCHOOL",
+  SPECIALIZED_SCHOOL: "SPECIALIZED_SCHOOL",
+  APPRENTICESHIP: "APPRENTICESHIP",
+  EMPLOYEE: "EMPLOYEE",
+  INDEPENDANT: "INDEPENDANT",
+  SELF_EMPLOYED: "SELF_EMPLOYED",
+  ADAPTED_COMPANY: "ADAPTED_COMPANY",
+  POLE_EMPLOI: "POLE_EMPLOI",
+  MISSION_LOCALE: "MISSION_LOCALE",
+  CAP_EMPLOI: "CAP_EMPLOI",
+  NOTHING: "NOTHING", // @todo find a better key --'
+};
+
+const STEPS = {
+  PROFIL: "PROFIL",
+  COORDONNEES: "COORDONNEES",
+  AVAILABILITY: "AVAILABILITY",
+  PARTICULIERES: "PARTICULIERES",
+  REPRESENTANTS: "REPRESENTANTS",
+  CONSENTEMENTS: "CONSENTEMENTS",
+  DOCUMENTS: "DOCUMENTS",
+  DONE: "DONE",
 };
 
 module.exports = {
@@ -603,8 +561,6 @@ module.exports = {
   updatePlacesSessionPhase1,
   updateCenterDependencies,
   deleteCenterDependencies,
-  assignNextYoungFromWaitingList,
-  sendAutoAffectationMail,
   updatePlacesBus,
   sendAutoCancelMeetingPoint,
   listFiles,
@@ -621,4 +577,7 @@ module.exports = {
   updateYoungStatusPhase2Contract,
   checkStatusContract,
   YOUNG_STATUS,
+  YOUNG_SITUATIONS,
+  STEPS,
+  inscriptionCheck,
 };